backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
aiofiles==23.2.1
google-generativeai==0.3.2
websockets==12.0
pillow==10.2.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
